+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [DEV] - 2023-04-06
+- Changed: refactored logic to load fieldpalette fields from dca and add to database
+- Removed: huh.fieldpalette.listener.callback service alias
+
+
 ## [0.6.13] - 2022-11-30
 - Changed: better error message
 - Fixed: recursion (Issue: [#16], PR: [#17])